Moved popup location
